ai_backend/pyproject.toml
[project]
name = "ai-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "0-kodiya-0",email = "sanithujayafiverr@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "huggingface-hub (>=0.29.3,<0.30.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "asgiref (>=3.8.1,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
